Combined tasks 02

+++ src/application/usecases/workout/delete-workout-usecase.ts
@@ -1,9 +1,14 @@
-import { IWorkout } from "@domain/entities/workout.entity";
+import { Workout } from "@domain/entities/workout.entity";
 import { IWorkoutRepository } from "@domain/interfaces/IWorkoutRepository";
-import { validationError } from "@presentation/middlewares/error.middleware";
+import {
+  InternalServerError,
+  NotFoundError,
+  validationError,
+} from "@presentation/middlewares/error.middleware";
 import { WorkoutStatus } from "@shared/constants/index.constants";
 import { TYPES_REPOSITORIES } from "@di/types-repositories";
 import { injectable, inject } from "inversify";
+import { IDeleteWorkoutUC } from "@application/interfaces/usecases/IWorkoutUC";
 
 /**
  * Purpose: Handle the deletion of a specific workout set by its ID.
@@ -13,16 +18,21 @@ import { injectable, inject } from "inversify";
  */
 
 @injectable()
-export class DeleteWorkoutUseCase {
+export class DeleteWorkoutUseCase implements IDeleteWorkoutUC {
   constructor(
     @inject(TYPES_REPOSITORIES.WorkoutRepository)
     private workoutRepository: IWorkoutRepository
   ) {}
-  
-  async execute(setId: string): Promise<IWorkout> {
+
+  async execute(setId: string): Promise<Workout> {
+    const workoutSet = await this.workoutRepository.findById(setId);
+
+    if (!workoutSet) {
+      throw new NotFoundError("workout not found");
+    }
     const deletedWorkoutSet = await this.workoutRepository.delete(setId);
     if (!deletedWorkoutSet) {
-      throw new validationError(WorkoutStatus.FailedToDelete);
+      throw new InternalServerError(WorkoutStatus.FailedToDelete);
     }
     return deletedWorkoutSet;
   }
